+++ csms/session.go
@@ -59,6 +59,8 @@ type Session struct {
 	handlerWG         sync.WaitGroup
 	readDone          chan struct{}
 	uniqueIDGenerator func() string
+	metrics           Metrics
+	metricSlots       chan struct{}
 	closed            chan struct{}
 	close             sync.Once
 	errMu             sync.RWMutex